create payment service WIP

+++ src/main/java/seasonlessrepayments/repository/RepaymentRepository.java
@@ -3,5 +3,5 @@ package seasonlessrepayments.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import seasonlessrepayments.model.Repayments;
 
-public interface RepaymentRepository extends JpaRepository<Repayments, Integer> {
+public interface RepaymentRepository extends JpaRepository<Repayments, Long> {
 }
